test: generalize e2e session test to run against both local cluster and Rotsee
Extract the shared session-pump harness from edgli_session_e2e.rs into
tests/common/mod.rs. Both tests now call run_one_megabyte_session_test(network)
with a Network::Local or Network::Rotsee variant. Rotsee provisioning reads
five EDGLI_ROTSEE_* env vars and requires no spawned process. Session targets
are discovered dynamically via connected_peer_addresses() instead of being
hardcoded as summary.nodes[0]/[2].
